% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_ds_json}
\alias{read_ds_json}
\title{read_ds_json}
\usage{
read_ds_json(ds_json_file)
}
\arguments{
\item{ds_jsonfile}{raw dataset json file}
}
\value{
tibble with the ds json object

This function takes a json objects and returns a tibbles

Function is mainly taken from i-akiya/R4DSJSON project
see: https://github.com/i-akiya/R4DSJSON
Licence: MIT
}
\description{
read_ds_json
}
