% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runViewer}
\alias{runViewer}
\title{shinyDSJSON}
\usage{
runViewer()
}
\value{
shiny.appobj object
}
\description{
shinyDSJSON
}
\examples{
library(devtools)
install_github("MichelLutz1994/shineDSJSON")

shineDSJSON::shinyDSJSON()

}
